parser/requirements.txt
requests
beautifulsoup4
mysql-connector-python
lxml
datetime
fake_useragent
pymysql
mysql.connector
